+++ app/src/main/res/layout/activity_extended_fab.xml
@@ -5,7 +5,7 @@
     android:id="@+id/rootLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ExtendedFabActivity">
+    tools:context=".components.fab.ExtendedFabActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
