+++ app/src/main/res/layout/activity_fab.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.FabActivity">
+    tools:context=".components.fab.FabActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
